Chapter 1.2 Problem 34
Written for Octave, may work with matlab.
There might be a typo in the book.  The problem states that for 3-d vectors the answer will be 2/pi but I banged my head for some time until I was helped online here:https://math.stackexchange.com/questions/2951802/average-size-of-dot-products-textbook-problem , the answer fits nicely with 2-d vectors which is what is included in the program.